Remove ottlconfig package
