--- pkg/ottl/ottlconfig/config.go
@@ -1,27 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ottlconfig // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/ottlconfig"
-
-// Config configures the OTTL statements to execute against traces, metrics, and logs.
-type Config struct {
-	Traces  SignalConfig `mapstructure:"traces"`
-	Metrics SignalConfig `mapstructure:"metrics"`
-	Logs    SignalConfig `mapstructure:"logs"`
-}
-
-// SignalConfig configures OTTL statements to execute.
-type SignalConfig struct {
-	Statements []string `mapstructure:"statements"`
-}
