+++ processor/transformprocessor/config_test.go
@@ -22,8 +22,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/confmap/confmaptest"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/ottlconfig"
 )
 
 func TestLoadConfig(t *testing.T) {
@@ -38,20 +36,20 @@ func TestLoadConfig(t *testing.T) {
 			id: config.NewComponentIDWithName(typeStr, ""),
 			expected: &Config{
 				ProcessorSettings: config.NewProcessorSettings(config.NewComponentID(typeStr)),
-				Config: ottlconfig.Config{
-					Traces: ottlconfig.SignalConfig{
+				OTTLConfig: OTTLConfig{
+					Traces: SignalConfig{
 						Statements: []string{
 							`set(name, "bear") where attributes["http.path"] == "/animal"`,
 							`keep_keys(attributes, "http.method", "http.path")`,
 						},
 					},
-					Metrics: ottlconfig.SignalConfig{
+					Metrics: SignalConfig{
 						Statements: []string{
 							`set(metric.name, "bear") where attributes["http.path"] == "/animal"`,
 							`keep_keys(attributes, "http.method", "http.path")`,
 						},
 					},
-					Logs: ottlconfig.SignalConfig{
+					Logs: SignalConfig{
 						Statements: []string{
 							`set(body, "bear") where attributes["http.path"] == "/animal"`,
 							`keep_keys(attributes, "http.method", "http.path")`,
