+++ processor/transformprocessor/factory_test.go
@@ -26,8 +26,6 @@ import (
 	"go.opentelemetry.io/collector/pdata/plog"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.opentelemetry.io/collector/pdata/ptrace"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/ottlconfig"
 )
 
 func TestFactory_Type(t *testing.T) {
@@ -40,14 +38,14 @@ func TestFactory_CreateDefaultConfig(t *testing.T) {
 	cfg := factory.CreateDefaultConfig()
 	assert.Equal(t, cfg, &Config{
 		ProcessorSettings: config.NewProcessorSettings(config.NewComponentID(typeStr)),
-		Config: ottlconfig.Config{
-			Traces: ottlconfig.SignalConfig{
+		OTTLConfig: OTTLConfig{
+			Traces: SignalConfig{
 				Statements: []string{},
 			},
-			Metrics: ottlconfig.SignalConfig{
+			Metrics: SignalConfig{
 				Statements: []string{},
 			},
-			Logs: ottlconfig.SignalConfig{
+			Logs: SignalConfig{
 				Statements: []string{},
 			},
 		},
